refactor(#3479): extract turn_finalizer completion-signal/delivery-lease/watcher-backstop submodules
Phase-2 rank-9: behavior-preserving extraction of three cohesive leaf
clusters out of src/services/discord/turn_finalizer.rs into new child
modules. PURE MOVE — zero logic change.

- turn_finalizer/completion_signal.rs: `CompletionSignal` enum + the pure
  `completion_signal_from_transcript()` derivation + their 11 unit tests.
- turn_finalizer/delivery_lease.rs: the dormant `DeliveryLeaseCell` handler
  wrappers (`handle_acquire/commit/release_delivery`) + their 15 state-machine
  tests.
- turn_finalizer/watcher_backstop.rs: the watcher far-backstop tunables
  (`WATCHER_BACKSTOP_TERMINAL_PROBE_INTERVAL`/`_STREAK`) + the terminal-or-defer
  verdict pair (`watcher_backstop_turn_is_terminal`/`_signal_is_terminal`) +
  the pure signal-truth-table test.

Each moved item is re-imported into the root (`use self::<mod>::{...}`) so the
actor-loop / reconciler / `completion_signal_state` call sites stay
byte-identical. The only source deltas in moved code are two relocation path
re-qualifications (`super::tmux::` / `super::inflight::` →
`crate::services::discord::{tmux,inflight}::`), which resolve to the identical
items. Tests coupled to root-private helpers (`with_isolated_runtime_root`,
`make_shared_data_for_tests_with_storage`) — the async watcher-backstop suite —
intentionally remain in root.

turn_finalizer.rs prod LoC 1528 -> 1337 (-191); the three new modules are 48 /
87 / 120 prod LoC, all well under the 1000-prod-LoC giant threshold (no new
registry/ratchet entry needed). Lowered the giant ratchet baseline to 1337 and
re-synced the change-surfaces.md / generated inventory numbers.

Gates: cargo build (no new warnings), cargo test -p agentdesk turn_finalizer
(87 passed), cargo clippy --lib (no new warnings), cargo fmt --check,
audit_maintainability.py --check (0), ci-script-checks.sh (0).
