+++ ARCHITECTURE.md
@@ -544,7 +544,10 @@ src/
 │   │   │   ├── watcher_handoff.rs
 │   │   │   └── watcher_orphan_cleanup.rs
 │   │   ├── turn_finalizer/
-│   │   │   └── cleanup.rs
+│   │   │   ├── cleanup.rs
+│   │   │   ├── completion_signal.rs
+│   │   │   ├── delivery_lease.rs
+│   │   │   └── watcher_backstop.rs
 │   │   ├── voice_barge_in/
 │   │   │   ├── final_result_playback.rs
 │   │   │   ├── foreground_decision.rs
