+++ src/services/discord/turn_finalizer.rs
@@ -34,8 +34,29 @@ use super::SharedData;
 use super::{DeliveryLeaseCell, LeaseHolder, LeaseOutcome};
 
 mod cleanup;
+mod completion_signal;
+mod delivery_lease;
+mod watcher_backstop;
 
 pub(in crate::services::discord) use cleanup::SyntheticClaimSnapshot;
+// #3479 r9: completion-signal enum + pure derivation extracted; re-exported so
+// the `completion_signal_state` method and the watcher-backstop re-check below
+// reference them unqualified, byte-identical.
+pub(in crate::services::discord) use self::completion_signal::{
+    CompletionSignal, completion_signal_from_transcript,
+};
+// #3479 r9: dormant delivery-lease handlers extracted to the child module; the
+// actor-loop call sites below reference them unqualified, byte-identical.
+#[allow(unused_imports)] // handlers are `#[cfg(unix)]`-conditional + dormant.
+use self::delivery_lease::{
+    handle_acquire_delivery, handle_commit_delivery, handle_release_delivery,
+};
+// #3479 r9: watcher far-backstop tunables + terminal-or-defer verdict pair
+// extracted; the `reconcile` loop references them unqualified, byte-identical.
+use self::watcher_backstop::{
+    WATCHER_BACKSTOP_TERMINAL_PROBE_INTERVAL, WATCHER_BACKSTOP_TERMINAL_STREAK,
+    watcher_backstop_turn_is_terminal,
+};
 
 /// How often the reconciler `Tick` fires to re-check deadline-armed
 /// gate-timeout entries and garbage-collect the ledger.
@@ -56,23 +77,6 @@ const GATE_BACKSTOP: Duration = Duration::from_secs(8);
 /// (`watcher_backstop_turn_is_terminal`) so a still-live turn is deferred.
 const WATCHER_REGISTER_BACKSTOP: Duration =
     Duration::from_secs(super::placeholder_sweeper::ABANDON_THRESHOLD_SECS);
-
-/// #3277 (Defect C) — proven-terminal FAST path for the watcher far-backstop.
-/// In the #3277 incident the handed-off turn was already PROVABLY complete
-/// (JSONL terminator on disk) while its watcher owner sat parked at transcript
-/// EOF, so no data-driven finalize ever fired and the channel stayed stranded
-/// for the full 1800s. The reconciler therefore PROBES watcher-owned Pending
-/// entries with the STRICT (`at_deadline = false`) form of
-/// `watcher_backstop_turn_is_terminal`: after
-/// `WATCHER_BACKSTOP_TERMINAL_STREAK` terminal probes this interval apart, the
-/// far deadline is pulled in to `GATE_BACKSTOP` for a third (still strict)
-/// confirmation before finalizing. A single non-terminal probe resets the
-/// streak (paused / paused-live / flapping turns keep the generous horizon).
-const WATCHER_BACKSTOP_TERMINAL_PROBE_INTERVAL: Duration = Duration::from_secs(15);
-
-/// Consecutive terminal probes required before the fast path pulls the
-/// watcher far-backstop deadline in (see above).
-const WATCHER_BACKSTOP_TERMINAL_STREAK: u8 = 2;
 
 /// TTL after which a `Finalized` ledger entry is garbage-collected so the
 /// ledger stays bounded while still suppressing a late double-submit.
@@ -332,22 +336,6 @@ pub(in crate::services::discord) enum FinalizeOutcome {
     /// with a bounded deadline; the reconciler will finalize when the
     /// precondition clears.
     Deferred,
-}
-
-/// #3016 S1: the structural completion signal from the provider's JSONL
-/// transcript, independent of the ledger.
-/// - `Done` — strict reverse-scan found a definitive terminator (Claude
-///   `result`/`system{...}`, Codex `turn.completed`): structurally over.
-/// - `PausedLive` — in-flight or inconclusive evidence; conservatively a live,
-///   paused turn.
-/// - `Unknown` — no structured on-disk JSONL turn state (LegacyTmuxWrapper /
-///   ProcessBackend / ClaudeEAdapter, or a non-JSONL provider).
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-// #3016 S3: wired into the watcher fresh-idle finalize decision.
-pub(in crate::services::discord) enum CompletionSignal {
-    Done,
-    PausedLive,
-    Unknown,
 }
 
 /// Ledger phase for a single turn. Owned solely by the actor task; the
@@ -1158,185 +1146,6 @@ async fn do_finalize(
         removed_token: finish.removed_token,
         has_pending: has_pending_after_voice,
         mailbox_online: finish.mailbox_online,
-    }
-}
-
-// #3041 §2-§3 — delivery-lease handlers: thin wrappers over the
-// `DeliveryLeaseCell` state machine (mod.rs), run in the actor task. DORMANT
-// after the R2 revert (the watcher works the cell INLINE); kept + unit-tested
-// for the sink/bridge wiring (P1-2..).
-
-/// CAS-acquire for `(key, [start,end))` on behalf of `holder`. #3041, dormant
-/// in the non-test build (the watcher acquires the cell directly, B4).
-#[allow(dead_code)] // #3041: AcquireDelivery actor arm dormant until sink/bridge wiring.
-fn handle_acquire_delivery(
-    lease: &DeliveryLeaseCell,
-    key: TurnKey,
-    holder: LeaseHolder,
-    start: u64,
-    end: u64,
-    deadline_ms: u64,
-) -> bool {
-    lease.try_acquire(key, holder, start, end, deadline_ms)
-}
-
-/// Three-way commit; full `(holder, key, [start,end))` mismatch = no-op. #3041
-/// P1-1: a successful `Delivered` commit advances the channel's
-/// `confirmed_end_offset` watermark to `end` (§5.2), gated on the lease having
-/// actually committed (so a rejected stale/duplicate commit never touches the
-/// offset) and via `advance_watcher_confirmed_end`'s monotonic CAS (never a
-/// double-advance). `NotDelivered`/`Unknown` never advance: an ambiguous
-/// terminal must not claim bytes as delivered.
-fn handle_commit_delivery(
-    lease: &DeliveryLeaseCell,
-    key: TurnKey,
-    holder: LeaseHolder,
-    start: u64,
-    end: u64,
-    outcome: LeaseOutcome,
-    provider: &ProviderKind,
-    tmux_session_name: &str,
-    shared: &SharedData,
-) -> bool {
-    let committed = lease.commit(holder, key, start, end, outcome);
-    // `mod tmux` is `#[cfg(unix)]`; non-unix commits the lease without an
-    // advance and consumes the otherwise-unused unix-only params.
-    #[cfg(unix)]
-    if committed && outcome == LeaseOutcome::Delivered {
-        super::tmux::advance_watcher_confirmed_end(
-            shared,
-            provider,
-            key.channel_id,
-            tmux_session_name,
-            end,
-            "src/services/discord/turn_finalizer.rs:commit_delivery_advance",
-        );
-    }
-    #[cfg(not(unix))]
-    let _ = (shared, provider, tmux_session_name);
-    committed
-}
-
-/// Compare-and-release; full `(holder, key, [start,end))` match only. #3041.
-#[allow(dead_code)] // #3041 P1-0: dormant, wired in P1-1..
-fn handle_release_delivery(
-    lease: &DeliveryLeaseCell,
-    key: TurnKey,
-    holder: LeaseHolder,
-    start: u64,
-    end: u64,
-) -> bool {
-    lease.release(holder, key, start, end)
-}
-
-/// #3016 S1 (PURE) — the structural completion signal derived solely from the
-/// provider's on-disk JSONL transcript. Shared by the public
-/// `completion_signal_state` (see it for the Done/PausedLive/Unknown
-/// rationale) and the reconciler's watcher-backstop liveness re-check.
-pub(in crate::services::discord) fn completion_signal_from_transcript(
-    provider: &ProviderKind,
-    runtime_kind: Option<crate::services::agent_protocol::RuntimeHandoffKind>,
-    transcript_path: &std::path::Path,
-) -> CompletionSignal {
-    if !crate::services::tui_turn_state::provider_runtime_has_structured_jsonl_turn_state(
-        provider,
-        runtime_kind,
-    ) {
-        return CompletionSignal::Unknown;
-    }
-    if crate::services::tui_turn_state::jsonl_turn_end_terminator_idle(provider, transcript_path) {
-        CompletionSignal::Done
-    } else {
-        CompletionSignal::PausedLive
-    }
-}
-
-/// #3016 phase-5a — the reconciler's terminal-or-defer verdict for a
-/// watcher-owned `register_start` Pending. `at_deadline == true` is the
-/// NATURAL 1800s far-backstop expiry; `false` (the #3277 fast-path probe AND
-/// the re-check of a fast-path-PULLED deadline, codex r1) stays STRICTLY
-/// transcript-proven. Never finalizes a legitimately long paused-live turn:
-///   * NO LIVE handle — absent (also under the inflight `tmux_session_name`
-///     re-key below: #3277 verify-1, a `claim_or_reuse_watcher` ReuseExisting
-///     dispatch registers under the OWNER channel only), `cancel` set, or
-///     `heartbeat_stale()` (#3268) → terminal ONLY at the natural deadline
-///     (nothing is left to drive the pane). The strict mode DEFERS: a watcher
-///     replace/reuse leaves the registry transiently absent/stale while the
-///     transcript still says busy — absence proves nothing about the TURN;
-///     dead/absent authority stays with the far horizon, never the fast path.
-///   * live-but-`paused` (a Discord turn took the session over) → defer.
-///   * else `watcher_backstop_signal_is_terminal` on the transcript: `Done`
-///     terminal; `PausedLive` defers; `Unknown` (non-JSONL runtime) consults
-///     the pane-ready fallback ONLY at the natural deadline.
-fn watcher_backstop_turn_is_terminal(
-    shared: &Arc<SharedData>,
-    channel_id: ChannelId,
-    provider: &ProviderKind,
-    at_deadline: bool,
-) -> bool {
-    let inflight_tmux = super::inflight::load_inflight_state(provider, channel_id.get())
-        .and_then(|state| state.tmux_session_name);
-    let (tmux_session_name, output_path, paused) = {
-        let handle = match inflight_tmux.as_deref() {
-            Some(tmux) => shared.tmux_watchers.by_tmux_session.get(tmux),
-            None => shared.tmux_watchers.get(&channel_id),
-        };
-        let Some(handle) = handle else {
-            return at_deadline;
-        };
-        if handle.cancel.load(std::sync::atomic::Ordering::Relaxed) || handle.heartbeat_stale() {
-            return at_deadline;
-        }
-        (
-            handle.tmux_session_name.clone(),
-            handle.output_path.clone(),
-            handle.paused.load(std::sync::atomic::Ordering::Acquire),
-        )
-        // dashmap `Ref` dropped here, BEFORE the (blocking) pane capture below.
-    };
-    if paused {
-        return false;
-    }
-    let runtime_kind =
-        crate::services::tui_prompt_dedupe::runtime_binding_for_tmux_session(&tmux_session_name)
-            .map(|binding| binding.runtime_kind)
-            .or_else(|| {
-                crate::services::tmux_common::resolve_tmux_runtime_kind_marker(&tmux_session_name)
-            });
-    watcher_backstop_signal_is_terminal(
-        completion_signal_from_transcript(
-            provider,
-            runtime_kind,
-            std::path::Path::new(&output_path),
-        ),
-        at_deadline,
-        || {
-            crate::services::tui_turn_state::pane_ready_fallback_allowed(provider, runtime_kind)
-                && crate::services::provider::tmux_session_ready_for_input(
-                    &tmux_session_name,
-                    provider,
-                )
-        },
-    )
-}
-
-/// #3277 verify-3 — the verdict over the transcript completion signal. The
-/// strict mode (`allow_pane_probe == false`: fast-path probe and pulled
-/// re-check) treats `Unknown` (non-JSONL Gemini / OpenCode / Qwen / legacy
-/// wrapper: no provable terminator) as NON-terminal: the synchronous
-/// pane-capture fallback can misread a dialog or a long silent stretch as
-/// idle, and probing it every 15s would amplify the old once-per-1800s
-/// exposure ~120× (and block the actor task). Only the NATURAL at-deadline
-/// re-check (`true`) consults `pane_ready` — lazily, only on `Unknown`.
-fn watcher_backstop_signal_is_terminal(
-    signal: CompletionSignal,
-    allow_pane_probe: bool,
-    pane_ready: impl FnOnce() -> bool,
-) -> bool {
-    match signal {
-        CompletionSignal::PausedLive => false,
-        CompletionSignal::Done => true,
-        CompletionSignal::Unknown => allow_pane_probe && pane_ready(),
     }
 }
 
@@ -3484,255 +3293,11 @@ mod tests {
     }
 
     // =======================================================================
-    // #3016 S1 — read-only completion-signal + watcher-pending probes.
+    // #3016 S1 — read-only watcher-pending probes. (#3479 r9: the
+    // completion-signal derivation tests moved to the `completion_signal`
+    // child module alongside the enum + pure fn they exercise.)
     // Additive, dead until S3/S4; these prove the read-only contract now.
     // =======================================================================
-
-    use crate::services::agent_protocol::RuntimeHandoffKind;
-
-    fn write_transcript(lines: &[&str]) -> tempfile::NamedTempFile {
-        let file = tempfile::NamedTempFile::new().unwrap();
-        std::fs::write(file.path(), lines.join("\n")).unwrap();
-        file
-    }
-
-    // (a) Claude transcript ending in a real terminator → Done.
-    #[test]
-    fn completion_signal_claude_terminator_is_done() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"user","message":{"content":"hi"}}"#,
-            r#"{"type":"assistant","message":{"content":[{"type":"text","text":"done"}]}}"#,
-            r#"{"type":"result","result":"done","session_id":"s"}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                file.path(),
-            ),
-            CompletionSignal::Done,
-        );
-    }
-
-    // (b) Claude transcript still streaming (no terminator) → PausedLive.
-    #[test]
-    fn completion_signal_claude_streaming_is_paused_live() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"user","message":{"content":"hi"}}"#,
-            r#"{"type":"assistant","message":{"content":[{"type":"text","text":"working"}]}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                file.path(),
-            ),
-            CompletionSignal::PausedLive,
-        );
-    }
-
-    // (b) Claude transcript whose latest line is a partial selector fragment
-    // after a terminator (a just-restarted turn) → PausedLive (the strict scan
-    // refuses to fall through a partial new envelope).
-    #[test]
-    fn completion_signal_claude_partial_after_terminator_is_paused_live() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"result","result":"done","session_id":"s"}"#,
-            r#"{"ty"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                file.path(),
-            ),
-            CompletionSignal::PausedLive,
-        );
-    }
-
-    // (a) Codex transcript ending in `turn.completed` → Done.
-    #[test]
-    fn completion_signal_codex_terminator_is_done() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"session_meta","payload":{"id":"s","cwd":"/repo"}}"#,
-            r#"{"type":"response_item","payload":{"type":"message","role":"assistant","content":[{"type":"output_text","text":"done"}]}}"#,
-            r#"{"type":"turn.completed","usage":{"input_tokens":5,"output_tokens":3}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Codex,
-                Some(RuntimeHandoffKind::CodexTui),
-                file.path(),
-            ),
-            CompletionSignal::Done,
-        );
-    }
-
-    // (b) Codex transcript mid-tool-call (no terminator) → PausedLive.
-    #[test]
-    fn completion_signal_codex_inflight_is_paused_live() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"session_meta","payload":{"id":"s","cwd":"/repo"}}"#,
-            r#"{"type":"response_item","payload":{"type":"function_call","name":"run_cmd","arguments":"{}","call_id":"c1"}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Codex,
-                Some(RuntimeHandoffKind::CodexTui),
-                file.path(),
-            ),
-            CompletionSignal::PausedLive,
-        );
-    }
-
-    // #3016 S3 (Concern 1): a COMPLETED Codex `agent_message` written right
-    // before a tool call is MID-TURN — the turn has not ended. The lenient drain
-    // probe would call this Idle, but the finalize `Done` decision uses the
-    // turn-END-only probe, so it must resolve to PausedLive (NOT Done) and the
-    // watcher therefore CANNOT over-finalize the live turn.
-    #[test]
-    fn completion_signal_codex_completed_agent_message_is_paused_live_not_done() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"session_meta","payload":{"id":"s","cwd":"/repo"}}"#,
-            r#"{"type":"item.completed","item":{"type":"agent_message","text":"on it, running a tool next"}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Codex,
-                Some(RuntimeHandoffKind::CodexTui),
-                file.path(),
-            ),
-            CompletionSignal::PausedLive,
-            "a completed agent_message with no turn.completed is mid-turn → not Done",
-        );
-    }
-
-    // #3016 S3 (Concern 1): a Codex `event_msg{task_complete}` (a task signal,
-    // not the turn record) is likewise NOT the turn terminator → PausedLive.
-    #[test]
-    fn completion_signal_codex_task_complete_is_paused_live_not_done() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"session_meta","payload":{"id":"s","cwd":"/repo"}}"#,
-            r#"{"type":"event_msg","payload":{"type":"task_complete"}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Codex,
-                Some(RuntimeHandoffKind::CodexTui),
-                file.path(),
-            ),
-            CompletionSignal::PausedLive,
-        );
-    }
-
-    // #3016 S3 (Concern 1): a Claude mid-turn assistant message (no terminator)
-    // → PausedLive; and a Claude `system{init}` (session-start, not turn-end) is
-    // at-rest to the drain probe but must NOT be Done for the finalize decision.
-    #[test]
-    fn completion_signal_claude_init_and_mid_turn_are_paused_live_not_done() {
-        let fin = TurnFinalizer::spawn();
-        let mid_turn = write_transcript(&[
-            r#"{"type":"user","message":{"content":"hi"}}"#,
-            r#"{"type":"assistant","message":{"content":[{"type":"text","text":"thinking"}]}}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                mid_turn.path(),
-            ),
-            CompletionSignal::PausedLive,
-        );
-
-        let init_only =
-            write_transcript(&[r#"{"type":"system","subtype":"init","session_id":"s"}"#]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                init_only.path(),
-            ),
-            CompletionSignal::PausedLive,
-            "system{{init}} is a session-start marker, not a turn-end terminator → not Done",
-        );
-    }
-
-    // #3016 S3 (Concern 1): a Claude `system{turn_duration}` IS a real turn-end
-    // terminator → Done (the stricter probe still accepts the genuine
-    // system-family turn boundary, not only `result`).
-    #[test]
-    fn completion_signal_claude_turn_duration_is_done() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[
-            r#"{"type":"user","message":{"content":"hi"}}"#,
-            r#"{"type":"assistant","message":{"content":[{"type":"text","text":"done"}]}}"#,
-            r#"{"type":"system","subtype":"turn_duration","session_id":"s"}"#,
-        ]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                file.path(),
-            ),
-            CompletionSignal::Done,
-        );
-    }
-
-    // (c) Non-JSONL runtime (LegacyTmuxWrapper) → Unknown even with a terminator
-    // on disk: the probe must not speak to completion for a runtime that has no
-    // structured on-disk turn state.
-    #[test]
-    fn completion_signal_non_jsonl_runtime_is_unknown() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[r#"{"type":"result","result":"done","session_id":"s"}"#]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::LegacyTmuxWrapper),
-                file.path(),
-            ),
-            CompletionSignal::Unknown,
-        );
-        // ProcessBackend and ClaudeEAdapter are also non-JSONL.
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ProcessBackend),
-                file.path(),
-            ),
-            CompletionSignal::Unknown,
-        );
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Claude,
-                Some(RuntimeHandoffKind::ClaudeEAdapter),
-                file.path(),
-            ),
-            CompletionSignal::Unknown,
-        );
-    }
-
-    // (c) Non-JSONL PROVIDER (Qwen) → Unknown regardless of runtime kind.
-    #[test]
-    fn completion_signal_non_jsonl_provider_is_unknown() {
-        let fin = TurnFinalizer::spawn();
-        let file = write_transcript(&[r#"{"type":"result","result":"done","session_id":"s"}"#]);
-        assert_eq!(
-            fin.completion_signal_state(
-                &ProviderKind::Qwen,
-                Some(RuntimeHandoffKind::ClaudeTui),
-                file.path(),
-            ),
-            CompletionSignal::Unknown,
-        );
-    }
 
     // has_live_watcher_pending: true for a watcher-owned non-finalized entry.
     #[tokio::test(flavor = "current_thread", start_paused = true)]
@@ -4198,375 +3763,6 @@ mod tests {
         fn lease_cell_is_send_sync() {
             let c: Arc<DeliveryLeaseCell> = Arc::new(DeliveryLeaseCell::new(ChannelId::new(9)));
             _assert_send(&c);
-        }
-    }
-
-    // =======================================================================
-    // #3041 §2-§3 §6 P1-0 — Dormant `DeliveryLeaseCell` state-machine tests.
-    //
-    // The cell is wired into no call path yet (P1-1..), but its transitions
-    // are proven correct now: single-winner CAS acquire, three-way commit,
-    // compare-and-release no-op on holder mismatch, and deadline reclaim. The
-    // tests drive the cell directly (and through the dormant handler wrappers)
-    // because that is the logic later phases depend on.
-    // =======================================================================
-    mod delivery_lease {
-        use super::super::{
-            TurnKey, handle_acquire_delivery, handle_commit_delivery, handle_release_delivery,
-        };
-        use crate::services::discord::{
-            DeliveryLeaseCell, LeaseHolder, LeaseOutcome, LeaseSnapshot,
-        };
-        use serenity::model::id::ChannelId;
-        use std::sync::Arc;
-
-        fn cell() -> DeliveryLeaseCell {
-            DeliveryLeaseCell::new(ChannelId::new(42))
-        }
-
-        fn turn() -> TurnKey {
-            TurnKey::new(ChannelId::new(42), 7, 0)
-        }
-
-        #[test]
-        fn fresh_cell_is_unleased() {
-            let c = cell();
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-            assert_eq!(c.channel_id(), ChannelId::new(42));
-        }
-
-        #[test]
-        fn acquire_records_holder_range_and_deadline() {
-            let c = cell();
-            let h = LeaseHolder::Watcher { instance_id: 1 };
-            assert!(c.try_acquire(turn(), h, 10, 20, 1_000));
-            match c.read() {
-                LeaseSnapshot::Leased {
-                    holder,
-                    turn,
-                    deadline_ms,
-                    start,
-                    end,
-                } => {
-                    assert_eq!(holder, h);
-                    assert_eq!(turn.exact_key(), self::turn().exact_key());
-                    assert_eq!(deadline_ms, 1_000);
-                    assert_eq!((start, end), (10, 20));
-                }
-                other => panic!("expected Leased, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn acquire_cas_admits_a_single_winner() {
-            // Two distinct holders race to acquire the SAME fresh cell; exactly
-            // one wins the CAS and the loser is rejected without mutating state.
-            let c = cell();
-            let w1 = LeaseHolder::Watcher { instance_id: 1 };
-            let w2 = LeaseHolder::Watcher { instance_id: 2 };
-            assert!(c.try_acquire(turn(), w1, 0, 5, 1_000));
-            // Second acquire on an already-Leased cell loses.
-            assert!(!c.try_acquire(turn(), w2, 0, 5, 1_000));
-            // The winner's payload is intact (loser did not overwrite it).
-            match c.read() {
-                LeaseSnapshot::Leased { holder, .. } => assert_eq!(holder, w1),
-                other => panic!("expected Leased held by winner, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn concurrent_acquire_has_exactly_one_winner() {
-            // Stronger single-winner proof: spawn N threads contending on one
-            // shared cell; exactly one try_acquire returns true.
-            use std::sync::atomic::{AtomicUsize, Ordering};
-            let c = Arc::new(cell());
-            let wins = Arc::new(AtomicUsize::new(0));
-            let barrier = Arc::new(std::sync::Barrier::new(16));
-            let mut handles = Vec::new();
-            for i in 0..16u64 {
-                let c = Arc::clone(&c);
-                let wins = Arc::clone(&wins);
-                let barrier = Arc::clone(&barrier);
-                handles.push(std::thread::spawn(move || {
-                    barrier.wait();
-                    if c.try_acquire(turn(), LeaseHolder::Watcher { instance_id: i }, 0, 1, 9_999) {
-                        wins.fetch_add(1, Ordering::Relaxed);
-                    }
-                }));
-            }
-            for h in handles {
-                h.join().unwrap();
-            }
-            assert_eq!(wins.load(Ordering::Relaxed), 1, "CAS must admit one winner");
-            assert!(matches!(c.read(), LeaseSnapshot::Leased { .. }));
-        }
-
-        #[test]
-        fn commit_three_way_delivered_not_delivered_unknown() {
-            for outcome in [
-                LeaseOutcome::Delivered,
-                LeaseOutcome::NotDelivered,
-                LeaseOutcome::Unknown,
-            ] {
-                let c = cell();
-                let h = LeaseHolder::Sink;
-                assert!(c.try_acquire(turn(), h, 3, 9, 1_000));
-                assert!(
-                    c.commit(h, turn(), 3, 9, outcome),
-                    "holder may commit {outcome:?}"
-                );
-                match c.read() {
-                    LeaseSnapshot::Committed {
-                        holder,
-                        start,
-                        end,
-                        outcome: got,
-                        ..
-                    } => {
-                        assert_eq!(holder, h);
-                        assert_eq!((start, end), (3, 9));
-                        assert_eq!(got, outcome);
-                    }
-                    other => panic!("expected Committed({outcome:?}), got {other:?}"),
-                }
-            }
-        }
-
-        #[test]
-        fn commit_by_non_holder_is_noop() {
-            let c = cell();
-            let owner = LeaseHolder::Watcher { instance_id: 1 };
-            let other = LeaseHolder::Watcher { instance_id: 2 };
-            assert!(c.try_acquire(turn(), owner, 0, 4, 1_000));
-            // Holder mismatch: commit refused, state stays Leased.
-            assert!(!c.commit(other, turn(), 0, 4, LeaseOutcome::Delivered));
-            assert!(matches!(c.read(), LeaseSnapshot::Leased { .. }));
-        }
-
-        #[test]
-        fn commit_on_unleased_is_noop() {
-            let c = cell();
-            assert!(!c.commit(LeaseHolder::Bridge, turn(), 0, 1, LeaseOutcome::Delivered));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-        }
-
-        #[test]
-        fn release_compare_and_release_noop_on_holder_mismatch() {
-            let c = cell();
-            let owner = LeaseHolder::Bridge;
-            let stale = LeaseHolder::Watcher { instance_id: 99 };
-            assert!(c.try_acquire(turn(), owner, 0, 8, 1_000));
-            // A stale actor cannot release the live lease.
-            assert!(!c.release(stale, turn(), 0, 8));
-            assert!(matches!(c.read(), LeaseSnapshot::Leased { .. }));
-            // The true holder releases successfully → back to Unleased.
-            assert!(c.release(owner, turn(), 0, 8));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-        }
-
-        #[test]
-        fn release_after_commit_returns_to_unleased() {
-            let c = cell();
-            let h = LeaseHolder::Sink;
-            assert!(c.try_acquire(turn(), h, 0, 2, 1_000));
-            assert!(c.commit(h, turn(), 0, 2, LeaseOutcome::Delivered));
-            // Release is valid from Committed for the recorded holder.
-            assert!(c.release(h, turn(), 0, 2));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-            // Idempotent: a second release on the now-Unleased cell is a no-op.
-            assert!(!c.release(h, turn(), 0, 2));
-        }
-
-        #[test]
-        fn stale_turn_commit_and_release_are_noops_after_reacquire() {
-            // #3041 §2 hazard, closed: turn A is acquired then reclaimed; turn B
-            // reacquires the SAME channel with the SAME holder KIND. A stale
-            // commit OR release carrying turn A's key must be a NO-OP and must
-            // NOT touch turn B's live lease. (Holder kind alone would match —
-            // only the stored turn identity distinguishes the two.)
-            let c = cell();
-            let holder = LeaseHolder::Sink; // same holder kind across both turns
-            let turn_a = TurnKey::new(ChannelId::new(42), 100, 0);
-            let turn_b = TurnKey::new(ChannelId::new(42), 200, 0);
-
-            // Turn A acquires, then its deadline elapses and it is reclaimed.
-            assert!(c.try_acquire(turn_a, holder, 0, 5, 10));
-            assert!(c.reclaim_if_expired(10));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-
-            // Turn B reacquires the freed cell (same channel, same holder kind).
-            assert!(c.try_acquire(turn_b, holder, 5, 11, 1_000));
-
-            // Stale commit from turn A: identity mismatch → no-op, B untouched.
-            assert!(!c.commit(holder, turn_a, 5, 11, LeaseOutcome::Delivered));
-            assert!(!c.commit(holder, turn_a, 0, 5, LeaseOutcome::Delivered));
-            // Stale release from turn A: identity mismatch → no-op, B untouched.
-            assert!(!c.release(holder, turn_a, 0, 5));
-            match c.read() {
-                LeaseSnapshot::Leased {
-                    turn, start, end, ..
-                } => {
-                    assert_eq!(turn.exact_key(), turn_b.exact_key(), "B still holds");
-                    assert_eq!((start, end), (5, 11));
-                }
-                other => panic!("turn B lease must survive stale A ops, got {other:?}"),
-            }
-
-            // Turn B's own commit/release with its real key still work.
-            assert!(c.commit(holder, turn_b, 5, 11, LeaseOutcome::Delivered));
-            assert!(!c.release(holder, turn_a, 5, 11)); // stale release post-commit: no-op
-            assert!(c.release(holder, turn_b, 5, 11));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-        }
-
-        #[test]
-        fn same_turn_stale_range_release_is_noop_after_reacquire() {
-            // #3041 codex R2: the SAME turn is reclaimed and reacquires a
-            // DIFFERENT byte range (e.g. a continuation chunk). A stale release
-            // carrying the OLD range — same holder AND same turn — must be a
-            // NO-OP and must NOT release the live newer-range lease. Only the
-            // correct range releases it (release is now range-scoped, symmetric
-            // with commit).
-            let c = cell();
-            let holder = LeaseHolder::Sink;
-            let t = TurnKey::new(ChannelId::new(7), 300, 0);
-
-            // Acquire range [0,5), let the deadline elapse, reclaim.
-            assert!(c.try_acquire(t, holder, 0, 5, 10));
-            assert!(c.reclaim_if_expired(10));
-            // Same turn reacquires a continuation range [5, 12).
-            assert!(c.try_acquire(t, holder, 5, 12, 1_000));
-
-            // Stale release with the OLD range [0,5): holder+turn match but the
-            // range does not → NO-OP, live [5,12) lease survives.
-            assert!(!c.release(holder, t, 0, 5));
-            match c.read() {
-                LeaseSnapshot::Leased { start, end, .. } => assert_eq!((start, end), (5, 12)),
-                other => {
-                    panic!("newer-range lease must survive stale-range release, got {other:?}")
-                }
-            }
-            // The correct range releases it.
-            assert!(c.release(holder, t, 5, 12));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-        }
-
-        #[test]
-        fn read_observes_payload_coherent_with_tag_under_race() {
-            // #3041 codex coherence fix: a reader that observes a non-`Unleased`
-            // state must observe the MATCHING payload — never a `Leased` tag
-            // paired with an `Unleased`/empty payload. Because `try_acquire`
-            // flips the tag AND writes the payload under one mutex (and `read`
-            // also locks), this holds by construction. Hammer it: while one
-            // thread repeatedly acquires/reclaims, readers must only ever see
-            // `Unleased` or a fully-populated `Leased{turn,range}` — never a
-            // torn intermediate.
-            use std::sync::atomic::{AtomicBool, Ordering};
-            let c = Arc::new(cell());
-            let stop = Arc::new(AtomicBool::new(false));
-            let t = turn();
-
-            let writer = {
-                let c = Arc::clone(&c);
-                let stop = Arc::clone(&stop);
-                std::thread::spawn(move || {
-                    while !stop.load(Ordering::Relaxed) {
-                        if c.try_acquire(t, LeaseHolder::Sink, 7, 13, 1) {
-                            // Immediately reclaim (deadline already in the past)
-                            // so the cell churns Unleased↔Leased rapidly.
-                            let _ = c.reclaim_if_expired(u64::MAX);
-                        }
-                    }
-                })
-            };
-
-            for _ in 0..200_000 {
-                match c.read() {
-                    LeaseSnapshot::Unleased => {}
-                    LeaseSnapshot::Leased {
-                        turn, start, end, ..
-                    } => {
-                        // The payload paired with the Leased state is always the
-                        // exact one the writer published — never torn/empty.
-                        assert_eq!(turn.exact_key(), t.exact_key());
-                        assert_eq!((start, end), (7, 13));
-                    }
-                    LeaseSnapshot::Committed { .. } => {
-                        panic!("writer never commits; tag/payload incoherent")
-                    }
-                }
-            }
-            stop.store(true, Ordering::Relaxed);
-            writer.join().unwrap();
-        }
-
-        #[test]
-        fn deadline_reclaim_forces_unleased_when_expired() {
-            let c = cell();
-            let h = LeaseHolder::Watcher { instance_id: 1 };
-            assert!(c.try_acquire(turn(), h, 0, 3, 100));
-            // Not yet expired: no reclaim.
-            assert!(!c.reclaim_if_expired(50));
-            assert!(matches!(c.read(), LeaseSnapshot::Leased { .. }));
-            // At/after the deadline: reclaimed regardless of holder identity.
-            assert!(c.reclaim_if_expired(100));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
-            // After a reclaim a fresh acquire can win again.
-            assert!(c.try_acquire(turn(), h, 0, 3, 200));
-        }
-
-        #[test]
-        fn deadline_reclaim_never_touches_committed() {
-            let c = cell();
-            let h = LeaseHolder::Bridge;
-            assert!(c.try_acquire(turn(), h, 0, 3, 10));
-            assert!(c.commit(h, turn(), 0, 3, LeaseOutcome::Delivered));
-            // A Committed lease awaits an explicit release; deadline reclaim is a
-            // no-op even far past the (now meaningless) deadline.
-            assert!(!c.reclaim_if_expired(10_000));
-            assert!(matches!(c.read(), LeaseSnapshot::Committed { .. }));
-        }
-
-        #[test]
-        fn dormant_handlers_drive_the_same_transitions() {
-            // The actor-task handler wrappers must produce identical results to
-            // the direct cell methods (they are wired in P1-1.. and exercised
-            // through these wrappers).
-            let c = cell();
-            let h = LeaseHolder::Watcher { instance_id: 3 };
-            assert!(handle_acquire_delivery(&c, turn(), h, 0, 6, 1_000));
-            assert!(!handle_acquire_delivery(
-                &c,
-                turn(),
-                LeaseHolder::Sink,
-                0,
-                6,
-                1_000
-            ));
-            // #3041 P1-1: the commit handler now takes provider/session/shared so
-            // a `Delivered` commit can advance the channel watermark. Supply a
-            // throwaway `SharedData`; the advance targets the cell's channel (42).
-            let shared = super::super::super::make_shared_data_for_tests_with_storage(None, None);
-            assert!(handle_commit_delivery(
-                &c,
-                turn(),
-                h,
-                0,
-                6,
-                LeaseOutcome::Delivered,
-                &crate::services::provider::ProviderKind::Claude,
-                "dormant-handler-test-session",
-                &shared,
-            ));
-            assert!(!handle_release_delivery(
-                &c,
-                turn(),
-                LeaseHolder::Watcher { instance_id: 4 },
-                0,
-                6
-            ));
-            assert!(handle_release_delivery(&c, turn(), h, 0, 6));
-            assert!(matches!(c.read(), LeaseSnapshot::Unleased));
         }
     }
 
@@ -5462,54 +4658,6 @@ mod tests {
             let _ = std::fs::remove_file(&transcript);
         })
         .await;
-    }
-
-    /// #3277 verify-3 (MINOR) truth table: the fast-path probe
-    /// (`allow_pane_probe == false`) must NEVER report `Unknown` (non-JSONL
-    /// runtime) as terminal — and must not even RUN the pane capture — while
-    /// the at-deadline re-check keeps the pane-ready fallback. `Done` /
-    /// `PausedLive` verdicts are identical in both modes.
-    #[test]
-    fn non_jsonl_signal_never_terminal_on_fast_path_probe() {
-        use std::cell::Cell;
-        // Unknown + fast path: non-terminal AND the pane capture must not run.
-        let captured = Cell::new(false);
-        assert!(!watcher_backstop_signal_is_terminal(
-            CompletionSignal::Unknown,
-            false,
-            || {
-                captured.set(true);
-                true
-            }
-        ));
-        assert!(
-            !captured.get(),
-            "the 15s fast-path probe must never run a blocking pane capture"
-        );
-        // Unknown + at-deadline: pane fallback decides (both directions).
-        assert!(watcher_backstop_signal_is_terminal(
-            CompletionSignal::Unknown,
-            true,
-            || true
-        ));
-        assert!(!watcher_backstop_signal_is_terminal(
-            CompletionSignal::Unknown,
-            true,
-            || false
-        ));
-        // Done / PausedLive: identical in both modes, pane never consulted.
-        for probe in [false, true] {
-            assert!(watcher_backstop_signal_is_terminal(
-                CompletionSignal::Done,
-                probe,
-                || unreachable!("Done must not consult the pane")
-            ));
-            assert!(!watcher_backstop_signal_is_terminal(
-                CompletionSignal::PausedLive,
-                probe,
-                || unreachable!("PausedLive must not consult the pane")
-            ));
-        }
     }
 
     /// #3277 verify-3 (MINOR): a watcher-owned turn on a non-JSONL runtime
